% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additor_sqlite_multiple_db.R
\name{additor_sqlite_multiple_db}
\alias{additor_sqlite_multiple_db}
\title{Write a RRD data table}
\usage{
additor_sqlite_multiple_db(input, output, tn_postfix = NULL)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data to}
}
\value{
returns invisible\code{TRUE}
}
\description{
Add new data to the database.
If a table exist, append it, if not,
create a new table with the name \code{name} if \code{create_new_table = TRUE}.
}
