% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additor_sqlite_multiple_db.R
\name{connect}
\alias{connect}
\title{Open connection to db}
\usage{
connect(dbname)
}
\arguments{
\item{dbname}{name of database}
}
\value{
connection object
}
\description{
Also create status report file in the dabname directory
}
