% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hash_exists.R
\name{db_hash_exists}
\alias{db_hash_exists}
\title{Check if hash already exists in table}
\usage{
db_hash_exists(table, hash)
}
\arguments{
\item{table}{tablename to check against (only one!)}

\item{hash}{hash to check (can be vector)}
}
\value{
logical vector of the same length as \code{hash}, \code{TRUE} if hash exists in table, otherwise FALSE
}
\description{
Check if hash already exists in table
}
