% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additor_sqlite_multiple_db.R
\name{db_name_from_fn}
\alias{db_name_from_fn}
\title{Get dbname from file name}
\usage{
db_name_from_fn(fn, dbname, seperate_db)
}
\arguments{
\item{fn}{file name}

\item{dbname}{default dbname}

\item{seperate_db}{}
}
\value{

}
\description{
Get dbname from file name
}
