% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additor_sqlite_multiple_db.R
\name{get_inputfiles}
\alias{get_inputfiles}
\title{Get file names of data files to be added to database}
\usage{
get_inputfiles(input, measure, include_files)
}
\arguments{
\item{input}{input folder}

\item{measure}{measure (folder name)}

\item{include_files}{list of files to be included}
}
\value{
file names to be imported
}
\description{
Get file names of data files to be added to database
}
