% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_DB.R
\name{initialize_db}
\alias{initialize_db}
\title{Create folder structure for data import and processing}
\usage{
initialize_db(config_file, tts_api_key = Sys.getenv("api_key"))
}
\arguments{
\item{config_file}{config file to use. If none is specified, \code{cofig.yml} in the current working directory will be used.}

\item{tts_api_key}{api key for getting the Trusted Timestamp from OriginStamp. Default is reading from the environmental variable \code{api_key}.}
}
\value{
invisible \code{TRUE}
}
\description{
Create folder structure for data import and processing
}
