% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_RRD.R
\name{new_RRD}
\alias{new_RRD}
\title{Create new RRD database}
\usage{
new_RRD(dbname)
}
\arguments{
\item{dbname}{the path and name of the database. Must not exist.}
}
\description{
Create a new database following the scheme used for the LEEF.RRD database
}
