% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_data.R
\name{remove_data}
\alias{remove_data}
\title{Remove records from the database}
\usage{
remove_data(db, timestamp, method)
}
\arguments{
\item{db}{name of the database from which the records should be removed}

\item{timestamp}{\bold{required argument!} timestamp which should be removed}

\item{method}{method(s) from which the records should be removed. This is the first string
before a double underscore (\code{__}) in the table name. If not specified, records with the timestamp
  are removed from all tables.}
}
\value{
name of the backup database or, if no table for the specified method exists, \code{NULL}.
}
\description{
Remove from the database all records from the database which are associated
  with a measurement method and have the provided timestamp. The records are
  moved to a backup database whose name is returned.
}
