% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_connect.R
\name{sqlite_connect}
\alias{sqlite_connect}
\title{Connect to raw data source}
\usage{
sqlite_connect()
}
\value{
invisibly the connection
}
\description{
The connection is saved in the option \code{data_connection}
}
