% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_disconnect.R
\name{sqlite_disconnect}
\alias{sqlite_disconnect}
\title{Disconnect from connected raw data source}
\usage{
sqlite_disconnect()
}
\value{
result from \code{DBI::dbDisconnect(conn)}
}
\description{
Disconnect from connected raw data source
}
