% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_isAlive.R
\name{sqlite_isAlive}
\alias{sqlite_isAlive}
\title{Check if connection to data source is alive}
\usage{
sqlite_isAlive()
}
\value{
\code{TRUE} if connection alive, otherwise \code{FALSE}
}
\description{
Checks if connection is alive and functional. If not, the function tries
to close the connection.
}
