% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additor_sqlite_multiple_db.R
\name{tn_from_fn}
\alias{tn_from_fn}
\title{Table name from file name and measure}
\usage{
tn_from_fn(fn, measure, tn_postfix)
}
\arguments{
\item{fn}{file name}

\item{measure}{measure}

\item{tn_postfix}{postfix for the table name}
}
\value{
table name
}
\description{
Table name from file name and measure
}
