% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additor_sqlite_multiple_db.R
\name{db_name_from_fn}
\alias{db_name_from_fn}
\title{Get dbname from file name}
\usage{
db_name_from_fn(fn, traits_fns, dbname)
}
\arguments{
\item{fn}{file name}

\item{traits_fns}{vector of files, which should be in the traits database}

\item{dbname}{default dbname}

\item{seperate_db}{}
}
\description{
Get dbname from file name
}
